Extract function to another file
